renamed component & small changes
