+++ modules/extensions/src/views/lite/components/debugger/inspectorTheme.tsx
