+++ modules/extensions/src/views/lite/components/debugger/style.scss
