+++ modules/extensions/src/views/lite/components/debugger/views/Entities.tsx
