+++ modules/extensions/src/views/lite/components/debugger/views/Inspector.tsx
