+++ modules/extensions/src/views/lite/components/debugger/views/Intents.tsx
