+++ modules/extensions/src/views/lite/components/debugger/views/Slots.tsx
