+++ modules/extensions/src/views/lite/components/debugger/views/Suggestions.jsx
