+++ modules/extensions/src/views/lite/index.jsx
@@ -1,3 +1,3 @@
 export { Dropdown } from './components/Dropdown'
-export { Wrapper } from './components/devtools/Wrapper'
-export { DevTools } from './components/devtools'
+export { Wrapper } from './components/debugger/Wrapper'
+export { Debugger } from './components/debugger'
