+++ modules/extensions/tsconfig.json
@@ -1,6 +1,7 @@
 {
   "compilerOptions": {
     "esModuleInterop": true,
-    "typeRoots": ["./src/botpress.d.ts", "./node_modules/@types", "./node_modules"]
+    "typeRoots": ["./src/botpress.d.ts", "./node_modules/@types", "./node_modules"],
+    "exclude": ["./src/views/**"]
   }
 }
